@@ -16,31 +16,32 @@
     }
   ],
   "require": {
-    "php": "^8.1",
-    "symfony/framework-bundle": "^5.4|^6.2",
-    "symfony/event-dispatcher": "^5.4|^6.2",
-    "symfony/property-info": "^5.4|^6.2",
-    "symfony/validator": "^5.4|^6.2",
-    "symfony/property-access": "^5.4|^6.2"
+    "php": "^8.3",
+    "symfony/framework-bundle": "^6.4|^7.3",
+    "symfony/event-dispatcher": "^6.4|^7.3",
+    "symfony/property-info": "^6.4|^7.3",
+    "symfony/validator": "^6.4|^7.3",
+    "symfony/property-access": "^6.4|^7.3"
   },
   "require-dev": {
     "ext-simplexml": "*",
-    "phpunit/phpunit": "^9.5",
+    "phpunit/phpunit": "^12",
     "friendsofphp/php-cs-fixer": "^3",
     "pedrotroller/php-cs-custom-fixer": "^2",
-    "vimeo/psalm": "^5",
-    "phpstan/phpstan": "^1",
-    "symfony/phpunit-bridge": "^5.4|^6.2",
-    "matthiasnoback/symfony-dependency-injection-test": "^4.3",
-    "nelmio/api-doc-bundle": "^4",
-    "symfony/stopwatch": "^5.4|^6.2",
-    "symfony/expression-language": "^5.4|^6.2",
-    "symfony/web-profiler-bundle": "^5.4|^6.2",
+    "vimeo/psalm": "^6",
+    "phpstan/phpstan": "^2",
+    "symfony/phpunit-bridge": "^6.4|^7.3",
+    "matthiasnoback/symfony-dependency-injection-test": "^6",
+    "nelmio/api-doc-bundle": "^5",
+    "symfony/stopwatch": "^6.4|^7.3",
+    "symfony/expression-language": "^6.4|^7.3",
+    "symfony/web-profiler-bundle": "^6.4|^7.3",
     "doctrine/doctrine-bundle": "^2",
-    "doctrine/orm": "^2|^3"
+    "doctrine/orm": "^2|^3",
+    "phpstan/phpstan-strict-rules": "^2.0"
   },
   "conflict": {
-    "nelmio/api-doc-bundle": "<4",
+    "nelmio/api-doc-bundle": "<5",
     "doctrine/doctrine-bundle": "<2.4"
   },
   "autoload": {
@@ -63,7 +64,7 @@
   },
   "extra": {
     "symfony": {
-      "require": "5.4.*",
+      "require": "6.4.*",
       "allow-contrib": false
     }
   }
